requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
PyPDF2==3.0.1
aiofiles==23.2.1
python-multipart==0.0.9
PyMuPDF==1.23.21  
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.10
translate==3.6.1
reportlab==4.2.0    
pandas==2.2.2       
python-docx==1.1.2  
python-pptx==0.6.23